Move kernel_info from the Kernel to the Shell.
